requirements.txt
Flask
tensorflow
opencv-python
numpy
Pillow
scikit-learn
